requirements.txt
httpx==0.27.2
openai==1.55.3
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.15
langchain-openai==0.2.9
faiss-cpu==1.9.0
pypdf==5.1.0
python-dotenv==1.0.1
streamlit==1.40.2
tiktoken==0.8.0